+++ src/pages/Lending/hooks/useGetEstimatedFeesQuery.ts
@@ -1,23 +1,12 @@
-import type { MarketData } from '@shapeshiftoss/types'
 import { useQuery } from '@tanstack/react-query'
 import { useMemo } from 'react'
 import type { EstimateFeesInput } from 'components/Modals/Send/utils'
 import { estimateFees } from 'components/Modals/Send/utils'
-import type { Asset } from 'lib/asset-service'
 import { bn } from 'lib/bignumber/bignumber'
 import { fromBaseUnit } from 'lib/math'
+import type { EstimatedFeesQueryKey } from 'lib/utils/thorchain/types'
 import { selectAssetById, selectMarketDataById } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
-
-export type EstimatedFeesQueryKey = [
-  'estimateFees',
-  {
-    enabled: boolean
-    asset: Asset | undefined
-    assetMarketData: MarketData
-    estimateFeesInput: EstimateFeesInput | undefined
-  },
-]
 
 // For use outside of react with queryClient.fetchQuery()
 export const queryFn = async ({ queryKey }: { queryKey: EstimatedFeesQueryKey }) => {
